Created separated test with separated rector
