+++ tests/PhpUnit/MultipleFilesChangedTrait/Fixture/fixture.php.inc
@@ -1,0 +1,19 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Rector\Core\Tests\PhpUnit\MultipleFilesChangedTrait\Fixture;
+
+final class MyClassName
+{
+
+}
+-----
+-----
+names.json
+-----
+-----
+{
+    "short": "MyClassName",
+    "fqn": "Rector\\Core\\Tests\\PhpUnit\\MultipleFilesChangedTrait\\Fixture\\MyClassName"
+}
