+++ tests/PhpUnit/MultipleFilesChangedTrait/Rector/CreateJsonWithNamesForClassRector.php
@@ -1,0 +1,43 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Rector\Core\Tests\PhpUnit\MultipleFilesChangedTrait\Rector;
+
+use Nette\Utils\Json;
+use PhpParser\Node;
+use PhpParser\Node\Stmt\Class_;
+use Rector\Core\Rector\AbstractRector;
+use Rector\FileSystemRector\ValueObject\AddedFileWithContent;
+use Symplify\RuleDocGenerator\ValueObject\RuleDefinition;
+
+final class CreateJsonWithNamesForClassRector extends AbstractRector
+{
+    public function getRuleDefinition(): RuleDefinition
+    {
+        return new RuleDefinition('Creates json with names for class', []);
+    }
+
+    public function getNodeTypes(): array
+    {
+        return [Class_::class];
+    }
+
+    /**
+     * @param Class_ $node
+     */
+    public function refactor(Node $node)
+    {
+        $smartFileInfo = $this->file->getSmartFileInfo();
+        $targetFilePath = $smartFileInfo->getRealPathDirectory() . '/names.json';
+
+        $content = Json::encode([
+            'short' => $this->nodeNameResolver->getShortName($node),
+            'fqn' => $this->getName($node),
+        ], Json::PRETTY);
+
+        $addedFileWithContent = new AddedFileWithContent($targetFilePath, $content);
+        $this->removedAndAddedFilesCollector->addAddedFile($addedFileWithContent);
+        return null;
+    }
+}
