+++ tests/PhpUnit/MultipleFilesChangedTrait/config/configured_rule.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-use Rector\Restoration\Rector\ClassLike\UpdateFileNameByClassNameFileSystemRector;
+use Rector\Core\Tests\PhpUnit\MultipleFilesChangedTrait\Rector\CreateJsonWithNamesForClassRector;
 use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
 
 return static function (ContainerConfigurator $containerConfigurator): void {
     $services = $containerConfigurator->services();
-    $services->set(UpdateFileNameByClassNameFileSystemRector::class);
+    $services->set(CreateJsonWithNamesForClassRector::class);
 };
